requirements.txt
flask==2.3.2
pandas==2.2.1
openpyxl==3.1.2

psycopg2-binary==2.9.9
